Revert "[*] 处理点分支问题"
This reverts commit 1f95ce3dd5b90265e67353d0bc31aab524fd42f2.
