+++ PowerThreadPool/Helpers/InterlockedFlag.cs
@@ -19,7 +19,7 @@ namespace PowerThreadPool.Helpers
 
         private string TypeName { get; } = typeof(T).Name;
 
-        internal string DebuggerDisplay => $"{TypeName}.{InterlockedValue}";
+        private string DebuggerDisplay => $"{TypeName}.{InterlockedValue}";
 
         private InterlockedFlag(T initialValue)
         {
@@ -49,10 +49,6 @@ namespace PowerThreadPool.Helpers
             if (ReferenceEquals(flag1, null))
             {
                 return ReferenceEquals(flag2, null);
-            }
-            else if (ReferenceEquals(flag2, null))
-            {
-                return ReferenceEquals(flag1, null);
             }
 
             return Interlocked.Read(ref flag1._innerValue) == Interlocked.Read(ref flag2._innerValue);
